+++ app/controllers/paperclip_upload/uploads_controller.rb
@@ -7,10 +7,18 @@ module PaperclipUpload
       respond_with PaperclipUpload::Upload.create(permitted_params), status: :created
     end
 
+    def download
+      send_file(upload.file.path, type: upload.file_content_type, disposition: 'inline')
+    end
+
     private
 
     def permitted_params
       params.permit(:file)
+    end
+
+    def upload
+      @upload ||= PaperclipUpload::Upload.find(params[:id])
     end
   end
 end