+++ app/serializers/paperclip_upload/upload_serializer.rb
@@ -1,3 +1,3 @@
 class PaperclipUpload::UploadSerializer < ActiveModel::Serializer
-  attributes :identifier
+  attributes :id, :identifier, :file_extension, :file_name
 end