+++ config/routes.rb
@@ -1,3 +1,5 @@
 PaperclipUpload::Engine.routes.draw do
-  resources :uploads, only: [:create], defaults: { format: :json }
+  resources :uploads, only: [:create], defaults: { format: :json } do
+    get 'download', on: :member
+  end
 end