+++ spec/dummy/app/controllers/api/base_controller.rb
@@ -1,0 +1,4 @@
+class Api::BaseController < ActionController::Base
+  self.responder = PaperclipUploadResponder
+  respond_to :json
+end