+++ spec/dummy/spec/controllers/paperclip_upload/uploads_controller_spec.rb
@@ -10,7 +10,21 @@ describe PaperclipUpload::UploadsController, type: :controller do
       it "returns 201" do
         post :create, file: file, format: :json
         expect(response).to be_success
+        expect(json_response.upload.id).not_to be_nil
         expect(json_response.upload.identifier).not_to be_nil
+        expect(json_response.upload.file_name).to eq('bukowski')
+        expect(json_response.upload.file_extension).to eq('jpg')
+      end
+    end
+  end
+
+  describe "GET 'download'" do
+    context "with defined upload" do
+      let!(:upload) { create(:upload) }
+
+      it "returns 200" do
+        get :download, id: upload.id, format: :json
+        expect(response).to be_success
       end
     end
   end